pyproject.toml
[project]
name = "council"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "boto3>=1.42.83",
    "chromadb>=1.5.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
